Conflicts: 	RentItService/Exceptions/InsufficientRightsException.cs 	RentItService/FunctionClasses/UploadDownload.cs 	RentItService/RentIt.Service.csproj 	RentItService/Services/ContentService.cs

+++ RentItService/Entities/Movie.cs
@@ -4,8 +4,6 @@
 // </copyright>
 //-------------------------------------------------------------------------------------------------
 
-using System.Collections.ObjectModel;
-
 namespace RentItService.Entities
 {
     using System;
@@ -64,6 +62,16 @@ namespace RentItService.Entities
         public string Genre { get; set; }
 
         /// <summary>
+        /// Gets or sets the owner ID.
+        /// </summary>
+        public int OwnerID { get; set; }
+
+        /// <summary>
+        /// Gets or sets the owner of the movie.
+        /// </summary>
+        public virtual User Owner { get; set; }
+
+        /// <summary>
         /// Gets or sets a list of rentals of the movie.
         /// </summary>
         public virtual ICollection<Rental> Rentals { get; set; }
@@ -80,19 +88,25 @@ namespace RentItService.Entities
         {
             Contract.Requires<ArgumentNullException>(token != null);
             Contract.Requires<ArgumentNullException>(movieObject != null);
-
-            Contract.Requires<InsufficientAccessLevelException>(User.GetByToken(token).Type == UserType.ContentProvider);
+            Contract.Requires<InsufficientRightsException>(User.GetByToken(token).Type == UserType.ContentProvider);
 
             using (var db = new RentItContext())
             {
-                foreach (var r in db.Rentals.Where(r => r.MovieID == movieObject.ID))
+                var movie = db.Movies.First(m => m.ID == movieObject.ID);
+
+                if (movie.OwnerID != User.GetByToken(token).ID)
+                {
+                    throw new InsufficientRightsException("Cannot delete a movie belonging to another content provider!");
+                }
+
+                foreach (var r in db.Rentals.Where(r => r.MovieID == movie.ID))
                 {
                     db.Rentals.Remove(r);
                 }
 
-                var filePath = Constants.UploadDownloadFileFolder + db.Movies.First(m => m.ID == movieObject.ID).FilePath;
+                var filePath = Constants.UploadDownloadFileFolder + movie.FilePath;
 
-                db.Movies.Remove(db.Movies.First(m => m.ID == movieObject.ID));
+                db.Movies.Remove(movie);
                 db.SaveChanges();
 
                 if (File.Exists(filePath))
@@ -112,7 +126,7 @@ namespace RentItService.Entities
         {
             Contract.Requires<ArgumentNullException>(token != null);
             Contract.Requires<ArgumentNullException>(search != null);
-            Contract.Requires<InsufficientAccessLevelException>(User.GetByToken(token).Type == UserType.SystemAdmin);
+            Contract.Requires<InsufficientRightsException>(User.GetByToken(token).Type == UserType.SystemAdmin);
 
             using (var db = new RentItContext())
             {
@@ -139,7 +153,7 @@ namespace RentItService.Entities
         {
             Contract.Requires<ArgumentNullException>(token != null);
             Contract.Requires<ArgumentNullException>(genre != null);
-            Contract.Requires<InsufficientAccessLevelException>(User.GetByToken(token).Type == UserType.SystemAdmin);
+            Contract.Requires<InsufficientRightsException>(User.GetByToken(token).Type == UserType.SystemAdmin);
 
             using (var db = new RentItContext())
             {
